feat(web): W2 S1 — extract MetaRecordFieldsPanel from MetaRecordDrawer (behavior-equivalent, frozen baseline)
design-lock: docs/development/multitable-w2-unified-record-inspector-design-lock-20260714.md §7 S1.

Extracts MetaRecordDrawer.vue's `details` tab body (pre-extraction L75-312: visibleFields loop +
all field editors + per-field AI shortcut + per-field comment anchor) into a new
MetaRecordFieldsPanel.vue child component. The panel receives the same props the details branch
consumed (record/fields/fieldPermissions/rowActions/commentPresence/link+person+attachment
summaries/uploadFn/deleteAttachmentFn/aiShortcut/buttonRunPending/mentionSuggestions) and re-emits
patch/ai-preview/ai-run/comment-field/open-link-picker/open-person-picker/run-button upward
UNCHANGED; the drawer relays each 1:1 to its own parent, so the workbench's existing listeners
need no changes. visibleFields (the field-mask filter) and canEditField move VERBATIM — masking is
load-bearing (history-field-mask class of bugs). formatValue/textControlValue/resolvedCanComment
are duplicated (not moved): the drawer's still-inline history tab and header comment-toggle still
need them; both copies are the same wrappers over the same shared utils. HI-1: the panel makes no
fetches of its own — zero new data paths.

Frozen-baseline discipline (P2-2a pattern): 24 specs that mount or mock MetaRecordDrawer (the 7
drawer-behavior specs the lock names, plus 6 field-editor specs that directly exercise the
extracted editors, plus 9 workbench-wiring specs, plus the comment-affordance CSS-consumer guard)
were captured pass-count-first, then re-run unmodified after extraction: 247/247 tests, identical
per-file counts on every file. The one deliberate non-baseline edit is comment-affordance-color-
consistency.spec.ts's CONSUMER_FILES list — the panel took the per-field comment anchor's
--active/--idle CSS rule with it (scoped styles must live with their template), so it is added to
keep that RATIFIED lock's guard covering the rule at its new home (drawer's own header comment-
active rule is untouched, so that guard already passed unmodified; this is additive coverage, not
a fix for a red).

New multitable-record-fields-panel.spec.ts (12 tests) covers the panel standalone: visibleFields
masking (a hidden field never renders), readOnly rendering, and all 7 emits with exact payload
shapes (patch, ai-preview/ai-run, comment-field incl. a negative control, open-link-picker vs
open-person-picker, run-button), plus a fetch-spy proving HI-1.

CI: wired `multitable-record-fields-panel` into run-required-web-tests.sh. Proved it executes
(335->336 files, 3907->3919 tests, exactly the new spec's 12) and bites (mutated one assertion ->
script exited 1 with the exact failing test; reverted via cp).

vue-tsc -b --noEmit clean throughout.
