+++ apps/web/tests/comment-affordance-color-consistency.spec.ts
@@ -28,6 +28,11 @@ const CONSUMER_FILES = [
   'src/multitable/components/MetaHierarchyView.vue',
   'src/multitable/components/MetaFormView.vue',
   'src/multitable/components/MetaRecordDrawer.vue',
+  // W2 S1 (2026-07-15): MetaRecordFieldsPanel.vue was extracted from MetaRecordDrawer.vue's
+  // `details` tab body and took the per-field comment anchor's `--active`/`--idle` rule with it
+  // (scoped CSS must live with the template that renders it). Added here so this lock keeps
+  // covering that rule at its new home instead of silently losing it off the guarded list.
+  'src/multitable/components/MetaRecordFieldsPanel.vue',
   'src/multitable/components/MetaCommentActionChip.vue',
   'src/multitable/components/MetaCommentAffordance.vue',
 ] as const
